requirements.txt
fastapi==0.115.12
supabase==2.15.0
python-dotenv==1.1.0
uvicorn==0.27.0
